In CLI mode, exiting process with 1 in case of fatal error or lint validation failure
To do this, I added fails to the instance object when outputting the failures. A callback is added
to the cli call to recess that computes the total number of errors across all instances. If either
an error is passed to that callback (file read failure, etc) or any of the instances have a fails
count > 1, then the process is exited with 1. (instead of 0 as default)

This addresses issue #23 and allows for much easier integration with Makefiles.