data-uri support as it presented at http://lesscss.org/#reference
"Inlines a resource and falls back to url() if the ieCompat option is on and the resource is too large, or if you use the function in the browser. If the mime is not given then node uses the mime package to determine the correct mime type.

Parameters:
mimetype: A mime type string. Optional.
url: The URL of the file to inline."

Added data-uri directive for any file to be embedded in css.
Also added option for Internet Explorer limitation of 32KB data-uri size. Available with "--dataUri ie"